backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
google-genai>=1.0.0
python-dotenv==1.0.1
chardet==5.2.0
